Restructuring cpp codes. Keep the bindings separately from the source code and as a single module.
