+++ c_src/utils.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <cmath>
 #include <random>
 #include <vector>
