+++ setup.cfg
@@ -22,7 +22,7 @@ classifiers =
 packages = ['sigpyproc']
 zip_safe = false
 include_package_data = false
-python_requires = ">=3.6"
+python_requires = >=3.6
 setup_requires =
     setuptools
     pybind11>=2.6.0
